Attempting cursor patch-back
When edits are made, the entire body of the document is rewritten. By
updating innerHTML we lose all DOM nodes, even ones that don't change,
so any information has to be recovered from whatever persists in the
Document model and whatever the Editor persists.

This attempts to use that strategy to determine the correct location of the cursor
after an edit by traversing the Document. It kinda works, the theory
seems solid, but the strategy is grating when contentEditable does this
work automatically.