Fix dynamic property response. Restructure js
attributeChangeCallback only fires for an attribute if its name is
returned by the static observedAttributes getter for a custom element
class. Attribute names are not case sensitive
and the API only sends events for element names in lower case.
The fix used here is to map all the strings to lower case.

Could replace the strings in hardcode, had thought to do it
dynamically, but decided to maintain the tiny mental abstraction for
developer benefit and preserve camel casing. The map works fine.

Can't dynamically update properly unless we use some metaprogramming to reach into
the class and change the getter which is a little dubious for this task.