Revert to strict zip, change margin fix approach
Instead of mixing the recursion fix by post-processing the data.
Test reveals that it doesn't do as well when a <p> has a single child
node of text, prepends a space. Going to try again by pre-processing now, with leading empty text bits
(so we get a "between" at the front and end)