Decent progress
Close correspondance now to offset and length of selection from start
More tight corners sadly.
Looks like we'll need to compute the 'true' offset into the selected,
and I was banking on that being perfect.
That or we need to enforce a strict HTML printing that favors the
selection algorithm