Update renderedText checker to fire when ready
If it fires immediately, the DOM hasn't loaded the content and we end up
with improperly formatted strings. That is, I think it ends up working
on the wrong content so there is a mismatch between the text we produce
from renderedText, and what the editor element has as its innerText.