Add function to get DOM selection from string offsets
Refactor Token to externalize the 'collapsing' functionality.
A list of tokens might need to be used for more than concatenating to
a string. A Token list is meant to have the contextual information to
know to to treat adjacent DOM elements. We use that to avoid burdening
Token clients with a cumbersome process.
Extracting it out gives us flexibility, which we use here to handle
Block elements differently.