Create more functional tooling for DOM work
As a note this is starting to attract a distinct "utilities module" code
smell that I want to avoid. I don't know if I can preempt it now but
a shift in responsibility should come at some point.

Again harness on capabilities.

Also try not to overuse the word capabilities