Initital sketch of interface for doc mutation
I say mutation only because the editor's perspective is the document
changes. Internally the document is immutable and 'mutations' are new
documents created from previous ones.

History will be tracked in a linked list for redos/undos.
Right now the DocHistory is in editor but it will be re-responsibilized
to Document