Refactor EditDocument to accommodate changes
With the change to handling Segments, EditDocuments have to be managed
a little differently. A pleasant side effect is much of the
functionality offloads to ListSegment

I might add that we do not have a test suite for EditDocument. The
refactoring *seems* to have gone smoothly.
I informally test it with the browser and my "test" page.
I'm pretty sure it's okay for now and the complexity of the project is
still small enough. We don't have reasons to put unit testing the DOM
interactions anymore since mocha supports in browser testing, and I
already have experience with cypress (and can probably pick up selenium)
but for now I'm testing components and those seem, to me, to be more
about final layouts. Guess it can do both.

Anyway, that's my note. The refactoring was smooth as far as I can tell.
I'd have more confidence in that assertion if we had tests.