Factor out 'rendering' capability
The interesting parts of the renders are the strings they produce, not
the action of displaying that string on a page. One of these is testable
(hint, it's harder to test that it's been put on a page.)
In related thinking, it would be prudent to isolate and narrow the
boundaries coupling DOM aspects to the capability suite. This makes it
easier to test (unit & integration) the capabilities without needing to
load it into a web page.
Though! We will need tests for that too, the important point is
separating concerns.