Lowkey change doc state changing
Had thought to make EditDocument immutable, but I'm feeling the
evolutionary forces lean toward EditDoc mutating. But in a way that the
content of the document is still immutable. In one sense, a "Document"
is a ListSegment. That is immutable and "changes" produce new
ListSegment. Then an "EditDocument" is a "current Document: and history
of "Document".
The "Edit" prefix derives from the fact that it tracks history of edits
and provides an API for causing those edits. Kind of like what Editor is
doing! Except Editor is the user facing interface, EditDocument is
internal, with Editor a kind of loose layer around that. One that
knows the appropriate time to, for example, re-render the DOM vs
letting ContentEditable handle it.