Include button to show off bug
Bug is that some <strong> tags nest and create differing "strengths" of
boldness. A motivator to work on tag collapsing for inline pieces.

Moreover another bug is ctrl-a. This, I think, selects the entire
content-editable, but doesn't work well with the document selection
mechanics. To resolve I think editor should shim in a piece that selects
within the document the whole thing, if the outer editor-hf (or, I
suppose, renderer. But the renderer would be nested in the editor-hf,
hm?). Yeah. Oof