Refactor EditDoc back to 'stateless' mutation
Will continue to pass back a new edition of the document.
As a consequence updated many functions to parameterize Document, where
before they'd use the 'this' reference. That strategy let's EditDocument
maintain an illusion of stateful updates. For example listeners attached
to a doc will get an event when the doc "changes", the doc doesn't
change but the listener is passed the new doc to operate on.