Fix tag duplication error
When rendering html and applying a style over a selection which already
contained that inline styling the resulting tags could nest, when we're
looking for a flat sequence.
The issue came because the process for updating the tags of a selection
would duplicate two tags that are the same because we use strings to
represent them. More fodder for my "String Problem" article