Remove empty segs for content change mutators
Content-preserving mutators (cut, split, copy) will leave empty segments
but the content updating ones don't. I do this mostly because these
functions all split, and split now produces a new segment each time even
if its empty. This is a challenge to work with because what is the
meaning to update an empty segment's tags? we could update the _locate*
functions to always make a minimal valid selection but that complicates
the semantics of _locate*. In a sense I'd like to preserve their
semantics. Additionally, if performing a lot of apply/remove/toggle tags
then the document begins tracking a myriad of useless segments.

Now! I could, of course, define ListSegment to be in an invalid state
if there are empty segments but I'm not keen to do that yet. Maybe if it
weren't for my own test case I might change it, but that is out of scope
I feel for this update. That would be a refactor for another purpose.