Add default attached to component
Have to work out semantics, for example if the component itself should
have the renderer (maybe a "rendererComponent" if we create wrapping
components for renderers") or if we might move it to Editor.
Editor has responsibility for manipulating the document and recalling
state. Knowing when to re-draw the document is out of purview, since
every renderer treats it differently.
There likely ought be determination in the hands of the renderer. It
would need comprehensive event interpreting to know when to redraw, and
in  such cases alert the editor to set selection.