Implement basic edit events to keep doc synced
For many edit actions the contentEditable attribute automatically
updates and renders what we want it to.
The trick is keeping the EditDoc in sync with the contents so that when
it renders back out to the DOM for the actions that contentEditable
doesn't cover we don't accidentally undo the actions of the user, as
the edit document serves as a source of truth.
If it's not up to date then it doesn't matter what the user sees on the
screen, it will get blasted away.