Non-committally edit out the test case
See. I'm thinking, the way it works now should be correct. I'm not
*strictly* fond of it. But, it's the behaviour of contentEditable and.
well. Far be it from *me* to disrupt a user experience.

generally, we ought take cues from contentEditable. So, that leaves me,
what to change?
Segments. It's a desync issue, technically, except a very "benign" one
or, sneaky one, that gives expected behavior.