Groundwork for my "Data System"
I'm thinking of it kinda like a DB wrapper. I want to abstract an
interface for saving and storing objects.
This I can hotplug into a configuration so long as it supports the
interface.

Right now it is a purely dev read/write over the filesystem and
data are held in memory, mostly.