Build API to use Dataccess
Rewire the frontend to properly 'slug' URLs
Not as keen on that because it low key forced me to use hypens in the
URLs too. Need to work on the querying capability, since querying
shouldn't have to match underlying data. But we should also not have to
query by the meta data (e.g id). That is WIP for now.