Update dataccess initialization
nextID field can be thought of as a "write head" for the DB. If we start
nextID in a place where incrementing it will reach another ID then we
would overwrite that memory.
The fix is to initialize it to start writing beyond the highest ID
available.
We can swap in a different identification scheme, or e.g vary per
Dataccess. This is a quick resolution for an incremental scheme.