Refactor move 'id' feature
Split handling. Database now now has custody of IDs and managing how
they are created.
Meanwhile Dataccess controls how IDs are accessed on an object.

Still work to be done here. The separation is not super clean. For
example Database relies on its client to set the ID when loaded.

We have this intermediate 'packed' data form that, likely, we should
not have. Clients may or may not wish to control how identifiers are
accessed but they shouldn't be responsible for actually accessing
them.
That is a TBD item I'm spending so much time on the data system and
not enough on the rest.