Restructure code to implement virtual indexing
"Virtual indexing" I mean. It's a basic method for mapping names to data
with the intent to keep IDs opaque.
Dataccess is the indeterrable representative for data and, for the time
being, seems the most natural place to start
that interface.

We have a big need to address. Querying data is not supported by my
persistence layer and I was instead creating a static map of references
to the data (the original 'notes' list).
This move brings me closer to a system for first-class handling of
arbitrary unstructured data. Having a simple index reduces some overhead
code to query each kind of stored data.