Fix updating the 'index' field of an object
Thing is: I don't know if I want this behaviour exposed.
An idempotency preference nags at me, that I should prevent updates to
the index. A repeated request to update has the same
effect, that is, no side effects. Or, each time the request is made the
outcome is the same, the end state is the same.
Right. I might. update it. again.