Update demo pages that reliad on removed feature
Editor components used to auto reformat the document. That was from a
time where I wanted to demonstrate the ability to work around whitespace
arbitrarily. Well, I succeeded. but decided that "white-space:
pre-wrap;" styling is a better fit for me needs.

However it is still useful as a capability to reformat HTML internally.
For example attaching an editor on a page with foreign HTML. We might
want to parse it as displayed rather than based on what a DOM stores
has internally.