Capture common copy pattern
Subclasses of section typically need to copy peticular state (that is
the main idea behind subclassing a Section). The default copy only
copies data relevant to a Section and it would be burdensome for a
subclass to override a method that uses copy only to also copy its
special state.
copyFrom takes care of that use case which lets us polymorphically copy
special state. The only requirement on a subclass is to override the
'copy' method. Then copyFrom will do the work of copying general Section
state and special subclass state.