Implement "tag" functionality to new Segment class
Update "map" to include "answer" functionality.
A Section can "answer" a function rather than propogate it. If answers,
then the section returns the result of applying the function to itself,
rather a clone of itself.
If a Section does not answer a function (default behavior) then it is
propogated as normal.