Update tests to drive change to 'operate'
Changing how 'operate' works so that it is more "structure preserving".
Informally it's closer to a map. An operate is a map that might split.
Currently it will split all the way up to the highest level in two
places for each layer. The change is that it only splits at the last
layer before the operation is applied.
If a layer can answer the operation, then the containing Section will
split and map over the chunks that can answer.