Fix mapRange implementation
It was, to put simply, not doing what I wanted.
The logic tree is a bit gruesome.
I would still like to do away with most of that.
I think it is possible to phrase this behavior the way I want, but
can't quite put my finger on how yet.

For now it is okay as long as it works.