Update "insert" to handle edge case
Also leave a note describing that edge case in detail.
Including a bit on my choice to use _locateAtom over _locateBoundary.
I like the notion that with _locateAtom I can collapse a set of edge
cases into one edge case. Solving that one solves the many.
It gives me desirable behavior (inserting at an out of bounds index,
when the index is too high, inserts at the end).
In a way, if you insert at the "length" you can imagine there is an
invisible "atom" at the end, like the null character of a string.
In fact I think that's how Segment handled it, because we stuck newlines
in some places, as a way of embedding <p> elements.

Well. It's good. might be a blog post in here.