First pass implementation of Doc delete
Add note about contextual ignorance and using locate boundary.
In particular we identify a need to know the "context" of which
segment a cursor is in when considering two cursor positions that are
"identical". A cursor that is placed between segments could indicate
either the left or right Section. We need to know which to properly
implement insert and delete.