Fix "straddle" behavior
"Straddle" behavior is maybe ill named. The idea is from the Doc
perspective a delete action that spans 2 or more Context should result
in the two contexts to either side of the delete space to be joined.
Liken it to a deleting from the middle of one paragraph up to the middle
of an adjacent one, or a later one.  The paragraphs join.
"Straddle behavior" implements those semantics.