Patch two of the failing asserts
Boundary aware deletes! Essentially: What happens when one or both of
start, end of a delete range are boundaries?
The "cutLeft" and "cutRight" are where only one is a boundary, and these
work now. The edge case is "cutMiddle" when both are.

For cut middle, the output should be an empty context. Trickily, I'd
like to avoid extensive logic checks to detect that situation.
I have a somewhat idea in progress to have this condition fall out
naturally but haven't had success yet, will keep trying.