Add "Gap" along with a bunch of comment notes
I kinda go off about the difficulties of "boundary" coordination as it
relates to Context. In the notes I resolved to try Gaps, which are an
idea I had while writing the notes.
I think I had a similar, if not the same idea, earlier but I didn't
commit it to writing then.