Fix some of the broken behavior with Gaps
Going with a "squeeze" notion. In this I imagine the gaps are more like
the ends of a vice, or a clamp.
Between them are 0 or more Context, then mergeTwoGaps closes the
vice/clamp, squeezing them together.

So... "Gap" isn't really the right word, more 'mark' or 'bookmark' or
'bookend'. Something.
And could likely be done more efficiently but I shan't optimize yet.