Fix incorrect document selections
When EditDoc used the Segment based internal document model we had to
compensate for extra newline charactars that Segments would contain for
indicating the end of a paragraph. Now that we are using the Section
model there is no need for the compensation. The Sections do not hold
any newline characters to represent paragrahps which caused the
selection to overshoot for selections after the first paragraph. The
overshoot error scaled linearly by adding 1 to the selected index per
paragraph after the first.