Implement new version of _locateBoundary
This one locates sections and offsets for boundaries!
We index over boundaries as we do for characters. Key idea is between
two Section there are two boundaries. The right most boundary for the
section on the left and the leftmost boundary for the section on the
right.
This concept will enable a little more control over special Section
behaviors. It will enable fine tuning what happens between Sections
by targeting boundaries.
It eleimates ambiguity from using standard character offsets.