Refactor move DOM dependency to EditDoc
Ultimately want to reduce the coupling of editor to DOM module.
At the same time its my preference that EditDoc know little about the
DOM. To me it suggests we upgrade the DOM module to control the
interface between DOM and EditDoc.

Then we can keep the DOM manipulating functions in one module and not
leak responsibility.