Refactor split to use boundaryIndices directly
rather than locating them from _locateAtomBoundary
TODO locateAtomBoundary is also not correctly implemented at the moment.
In my mind you give it an atom index and it can find the boundaries to
the left, right.
That is not necessarily difficult to do but I'm finding that, as I
slowly implement more and more of the Section functionality to use
boundaries, that the use cases for character indices diminish.