Update render to display "empty" Context
No context is "empty" by the function definition.
That only encourages me to change the name to something else.
In this case, the precise explanation is we are showing Context even if
they have no subPieces.
Why? because we can target them for insertion anyway.
A collapsed paragraph doesn't render, but in the DOM the output
of insert paragraph is to embed a <br>, so we do too.