Refine controls over how Section 'map' operates
Adds ability to distinguish "answers" behavior by a "targetedBy" 'flag'.
A section that answers but is not targetedBy will apply the function to
its children (default AtomicSection behavior).
A section that is targetedBy (implying "answers") will target /itself/
with the function.
In all "answers" cases the propogation stops. If neither targetedBy nor
answers then behavior recurses (default Section) behavior.

Currently there is no situation where targetedBy but not answers, this
is considered a contradiction (but is not enforced meaningfully).
Producing such a situation should not allow for a difference in
behavior. The requirement could be stated like so: If answers is false,
then the behavior for targetedBy is true and targetedBy is false should
be identical.