Fix bug causing mismatch between input and editDoc
Browser would insert <div> elements on enterNewLine.
This caused the cursor locating function to ignore the div elements and
return, as the cursor position, the last cursor position of the heading
element.
Note that when enterNewLine is used from a <p> element it would produce
a <p> element. Seems the bug was only detectable from the h1 element.

Ah.

The fix is to prevent the browser from doing whatever it wants and just
rendering what we have each time.
Likely should do this for ~any~ event.