Refactor Context to respect all boundaries
Previously overrode insertBoundary in Context to disregard boundaries
between Segments. This was a way to mimic the "cursor position" within
the document.
Instead now the insert works as it does for any Section.
It is incumbent on clients of Context, and Doc generally, to map the
cursor position to boundaries before calling insert methods.

May move the mapping logic to inside Doc, so that it exposes an
interface that accepts args as "cursor positions". Might do that.
For now letting EditDocument do the mapping.