Expand capability set of cursorToBoundary
Can now skip empty Segment when determining boundary position.
Effective outcome, when a cursor maps to a boundary it always favors
left. That is if a cursor can plausibly map to multiple boundaries it
would pick the "leftmost" one.

Now we can flip the preference to favor the "rightmost".
This let's us create minimal boundary selections for a given pair of
cursors.
The left favors the right, right favors left, bringing them as close as
possible when mapped to boundaries