Fix editDocument test
Logically I demand that new Contexts be paragraphs.
Rules written for a simpler world.
Now I need to be a little more nuanced.
I'd like a way to control for what should happen to each context along
the path toward the actual "break".
More to the point, I'd like to expose that to the Control engine, which
could express, then,  "set last two contexts as li, p"