Update tests to test for correct outputs
And in the process, now I see what deleteBoundary is. hm.
Let's say difficult to work with in the nested context.
There's a bigger re-arrangement that needds to happen.
I think it amounts to "popping contexts up to the next earliest
non-mixed Context as would come by in a pre-order traversal"

need some tree manipulating algorithms!
Leaf re-arrangement!