Swap out construction of 'component' for parsing
Want to favor parsed documents-as-test-casses for a few reasons.
One it makes me feel cool to use a DSL so there's that.
Two, comparing the pretty printed strings is a very quick way to
identify why actual !== expected. Mocha prints a nifty diff
when comparing strings, or arrays for example.
As opposed to "structure didn't match" forcing me to step through the
debugger.

Now.
Frequently I need to go in the debugger anyway.
But I sidestep a big initial knowledge cliff.
if a test fails first you have to identify what is failing and it is
not always a particularly easy task. This launches me into the know
 immediately.

Then once I DO step through the debugger I have an idea where to look.
Good stuff.