Implement marker hiding behavior for rendered Doc
Currently the styling is not enforced.
Thie change more or less *enables* devs to select the markers that
"should be hidden".

To that end likely I'd like to treat this classification more like a
pseudo class.
It indicates something about the state of a list item.
Here the state is "first child is a ul".

Then the styler can do what they'd like with that.