Serve dynamic pages for "app" data
Only problem with the static pages is they feel more appropriate for
publishing.
For the Alpha, for now, we'll assume that, for the "app", content will
be served dynamically and the pages will reflect that.

When the static site is built it won't have any server to reach out to.

TBD on what to do for that in the future.