Update login to always fetch user
The intent is to not reveal information about whether a user exists or
 not. In practice this could be accomplished anyway with a create user
page, which would reject attempts to create new accounts using an
existing username.

However for example when we integrate with emails we'll want to hide
whether an email is attached to an account or not.

I see this as practice, and some security mindset.