Incorporate escape sequences into Scribdoc
This is a fix of an earlier commit
So....
Token parsing, right. This likely would have been easier if I did char
by char lexing to work out if the amount of backslashses prior to a
single quote was a correct amount or not. like.
It would have needed some amount of look ahead/look behind anyway.

I am glad I figured it out with the regex but... that's a complicated RE